
<html>
<head>
<meta http-equiv="Content-Type" content="text/html; charset=utf-8"/>
</head>
<body>
Esto es una prueba con la ñ y los acentos ò


% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doble.R
\name{doble}
\alias{doble}
\title{Calcula el doble de un n&uacute;mero, es decir, lo multiplica por 2}
\usage{
doble(x)
}
\arguments{
\item{x}{número}
}
\value{
El doble de \code{x}
}
\description{
Calcula el doble de un número, es decir, lo multiplica por 2
}
\examples{
doble(7)
doble(8)
}
</body>
</html>
