% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummys.R
\name{multiplicando}
\alias{multiplicando}
\title{Multiplicar dos números}
\usage{
multiplicando(x, y)
}
\arguments{
\item{x}{e y números}
}
\description{
Esta es una funcion de ejemplo para multiplicar.
}
\details{
Esta funcion es dummy y no sirve para nada practicamente.....
}
