% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suma.dos.numeros.R
\name{suma.dos.numeros}
\alias{suma.dos.numeros}
\title{Sumatorio de dos numeros.}
\usage{
suma.dos.numeros(x, y)
}
\arguments{
\item{x}{un dato numerico.}

\item{y}{un dato numerico.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Sumatorio de dos numeros.
}
\examples{
suma.dos.numeros(12, 3)
suma.dos.numeros(10.8, 1.3)
}

