% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummys.R
\name{sumando}
\alias{sumando}
\title{sumando}
\usage{
sumando(x, y)
}
\arguments{
\item{x}{e y  numeros}
}
\description{
Esta es una funcion que tiene dos argumentos y los suma .
}
\details{
Esta funcion es dummy y no sirve para nada practicamente.....
}
